+++ src/init/features/dataconnect/sdk.ts
@@ -211,7 +211,7 @@ export async function generateSdkYaml(
   if (targetPlatform === Platform.IOS) {
     const swiftSdk = {
       outputDir: path.relative(connectorDir, path.join(appDir, `dataconnect-generated/swift`)),
-      package: upperFirst(camelCase(connectorYaml.connectorId)) + "Connector",
+      package: "DataConnectGenerated",
     };
     connectorYaml.generate.swiftSdk = swiftSdk;
   }
@@ -221,7 +221,7 @@ export async function generateSdkYaml(
     const packageJsonDir = path.relative(connectorDir, appDir);
     const javascriptSdk: JavascriptSDK = {
       outputDir: path.relative(connectorDir, path.join(appDir, `dataconnect-generated/js/${pkg}`)),
-      package: `@firebasegen/${pkg}`,
+      package: `@dataconnect/generated`,
       // If appDir has package.json, Emulator would add Generated JS SDK to `package.json`.
       // Otherwise, emulator would ignore it. Always add it here in case `package.json` is added later.
       // TODO: Explore other platforms that can be automatically installed. Dart? Android?
@@ -246,7 +246,7 @@ export async function generateSdkYaml(
         connectorDir,
         path.join(appDir, `dataconnect-generated/dart/${pkg}`),
       ),
-      package: pkg,
+      package: "dataconnect_generated",
     };
     connectorYaml.generate.dartSdk = dartSdk;
   }
@@ -254,7 +254,7 @@ export async function generateSdkYaml(
   if (targetPlatform === Platform.ANDROID) {
     const kotlinSdk: KotlinSDK = {
       outputDir: path.relative(connectorDir, path.join(appDir, `dataconnect-generated/kotlin`)),
-      package: `connectors.${snakeCase(connectorYaml.connectorId)}`,
+      package: `com.google.firebase.dataconnect.generated`,
     };
     // app/src/main/kotlin and app/src/main/java are conventional for Android,
     // but not required or enforced. If one of them is present (preferring the