+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added prefix support for multi-instance Cloud Functions extension parameters. (#8911)
 - Make it possible to init a dataconnect project in non interactive mode (#8993)
 - Added 2 new MCP tools for crashlytics `get_sample_crash_for_issue` and `get_issue_details` (#8995)
+- `firebase init dataconnect:sdk` generate consistent fixed package names (#9021)
 - Use Gemini to generate schema and seed_data.gql in `firebase init dataconnect` (#8988)
 - Fixed a bug when `firebase deploy --only dataconnect` didn't include GQL files in nested folders (#8981)
 - Changed `firebase deploy` create Cloud SQL instances asynchronously (#9004)